+++ frontend/components/ExamCard/index.tsx
@@ -2,7 +2,9 @@
 
 import './card.css';
 import { useState } from 'react';
-import { dummyFiles } from '../../lib/DummyData/dummyData';
+import { dummyFiles } from '@/lib/DummyData/dummyData';
+import { formatDate } from '@/lib/utils';
+import { downloadFile } from '@/lib/utils';
 
 export const ExamCard = () => {
   // 定義檔案的型別
@@ -14,9 +16,7 @@ export const ExamCard = () => {
     fileLocation: string;
   };
 
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const [isFirstList, setIsFirstList] = useState(true);
-  const [listName, setListName] = useState('Show Individual File');
+  const [currentIndex, setCurrentIndex] = useState(-1); // -1 表示顯示完整列表
 
   const handleNext = () => {
     setCurrentIndex((prevIndex) =>
@@ -31,28 +31,7 @@ export const ExamCard = () => {
   };
 
   const toggleList = () => {
-    setIsFirstList((prevState) => !prevState);
-    setListName(isFirstList ? 'Show All Files' : 'Show Individual File');
-  };
-
-  const downloadFile = (fileUrl: string, fileName: string) => {
-    const link = document.createElement('a');
-    link.href = fileUrl;
-    link.download = fileName;
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  };
-
-  const formatDate = (date: Date) => {
-    return date.toLocaleString('zh-TW', {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-      weekday: 'long',
-      hour: '2-digit',
-      minute: '2-digit',
-    });
+    setCurrentIndex((prevIndex) => (prevIndex === -1 ? 0 : -1));
   };
 
   // 將檔案資訊顯示邏輯提取成獨立的 FileInfo 子組件
@@ -77,20 +56,26 @@ export const ExamCard = () => {
     </div>
   );
 
-  const file = dummyFiles[currentIndex];
-
   return (
     <div className="card rounded-lg bg-white p-6 shadow-md">
       <h1 className="text-lg font-medium text-gray-800">Past Exam Files</h1>
       {/* 切換按鈕 */}
       <button className="toggle" onClick={toggleList}>
-        {listName}
+        {currentIndex === -1 ? 'Show Individual File' : 'Show All Files'}
       </button>
       {/* 檔案列表 */}
-      {!isFirstList && (
+      {currentIndex === -1 ? (
+        <ul className="text-black">
+          {dummyFiles.map((file) => (
+            <li className="card" key={file.filename}>
+              <FileInfo file={file} />
+            </li>
+          ))}
+        </ul>
+      ) : (
         <ul>
-          <li key={file.filename} className="text-black">
-            <FileInfo file={file} showDetails />
+          <li key={dummyFiles[currentIndex].filename} className="text-black">
+            <FileInfo file={dummyFiles[currentIndex]} showDetails />
             <div className="mt-4 flex justify-between">
               <button onClick={handlePrevious} className="button-switch">
                 Previous
@@ -100,15 +85,6 @@ export const ExamCard = () => {
               </button>
             </div>
           </li>
-        </ul>
-      )}
-      {isFirstList && (
-        <ul className="text-black">
-          {dummyFiles.map((file) => (
-            <li className="card" key={file.filename}>
-              <FileInfo file={file} />
-            </li>
-          ))}
         </ul>
       )}
     </div>
